requirements.txt
acres==0.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
deepseek==1.0.0
defusedxml==0.7.1
distro==1.9.0
etelemetry==0.3.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fitz==0.0.1.dev2
fqdn==1.5.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.8
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.25.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
openai==1.61.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
Send2Trash==1.8.3
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
streamlit==1.42.0
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
yarl==1.18.3
zstandard==0.23.0
